// routes/logic/classes/interfaces/PartnerInterface.ts
export interface PartnerInterface {
    partnerNumber: string;
    partnerPreName: string;
    partnerName: string;
    partnerCompany: string;
    partnerLanguage: string;
    partnerCreateDate: string;
    partnerEmail: string;
    partnerPhone: string;
    partnerType: string;
    partnerEecNumber: string;
    partnerEdiInvoice: string;
    partnerEdiOrderSp: string;
    partnerEdiDesadv: string;
    partnerEmailRg: string;
    partnerEmailLi: string;
    partnerEmailAu: string;
    partnerPhone0: string;
    partnerPhone1: string;
    partnerFax0: string;
    partnerMob0: string;
    partnerMob1: string;
    partnerCrnnum: string;
    partnerPaymentTermId: string;
}

export interface PartnerDataInterface {
    CUSTOMERS_NUMBER: string;
    CUSTOMERS_PRENAME: string;
    CUSTOMERS_NAME: string;
    CUSTOMERS_COMPANY: string;
    CUSTOMERS_TYPE: string;
    LANGUAGE: string;
    CREATE_DATE: string;
    CUSTOMERS_EMAIL: string;
    CUSTOMERS_PHONE: string;
    EEC_NUM: string;
    EDI_INVOIC: string;
    EDI_ORDERSP: string;
    EDI_DESADV: string;
    EMAIL_RG: string;
    EMAIL_LI: string;
    EMAIL_AU: string;
    PHONE_0: string;
    PHONE_1: string;
    FAX_0: string;
    MOB_0: string;
    MOB_1: string;
    CRNNUM: string;
    PAYMENT_TERM_ID: string;
}